Auth with Microsoft instead of Facebook
When user logs in, the MS access token is saved in the Redux store and saved to the device's file system. The token is also displayed on the "Me" screen, just for now, to "prove" that we've got the real thing. Of course, we would not show this in the final product.

To make this work, I added a "native" section to the [MS app configuration](https://apps.dev.microsoft.com). This thingy needed a redirect url that did not begin with the `https` scheme. Expo's `AuthSession` uses `https` scheme for it's redirect url by default, so I had to "hardcode" the redirect url into `config.js`. I'm really not sure if this is the way to go, but it seems to be working so far. The rubber will meet the road when we deploy to various "environments".