Use id token when making requests to Bookit API
Use the Microsoft id_token instead of access_token.

Pull user name out of token and display to the user on the "Me" page

Move all network requests out of the `actions` folder to an `api` folder.

Send the Microsoft id token in all calls to bookit-api.

Set up test framework for testing api layer. This will (hopefully) serve as a guideline for testing all code that makes a network request. The tests rely on a mocked version of `axios`, which allows us to test our api layer without actually making any network requests. (See `src/api/__mocks__`.) Within tests, calls made using `axios.get()` return an array of integers. Calls using `axios.post()` resolve to a mocked response as well.

Question: Does the mock need to return data that mimics the shape of the real data? I don't think I need to do this in order to test the functionality of the api layer, but maybe this is something we should discuss as a team.

Question: When the api functions are not called with the right params, they return a rejected promise. I'm wondering if they should throw instead? Along those lines, the test framework seems to prefer that rejected promises reject with a string (which seems weird) while ESLint prefers that promises reject with an error object. I disabled the ESLint rule, and we can revisit this if it becomes an issue.

Reference:
https://facebook.github.io/jest/docs/en/asynchronous.html
https://facebook.github.io/jest/docs/en/manual-mocks.html