Rough draft of our testing strategy
The intention is to provide a high level conceptual overview of our
testing strategy. Once this is in place (more or less) we can
prioritize various aspects of implementing the strategy, identify gaps
in our knowledge and toolset, and then, finally, get to writing the tests themselves!

Hopefully this will give context to metrics such as "code coverage". The
last thing we want to do, of course, is write tests merely for the sake
of bumping up an arbitrary number on somebody's dashboard :)

I'm basing this on Martin Fowler's pyramid. We could use anything -- I decided to start with this because it is a well-known and widely accepted conceptual framework.

This document is missing an explanation of the "Service" layer. We need
to decide what this means for us, and then decide how to actually implement it.