Remove unit test from pre-push validation
Unit tests now fail if coverage is too low. Coverage will be too low for a long while yet, which forces devs to do a `--no-verify` on pushes. Since we don't want to get folks in the habit of skipping verification, we're skipping the unit tests on pushes.

We will add unit tests to the pre-push check once we have meaningful unit tests.

Also added coverage for jsx, and removed coverage for typescript. (We're not using typescript!)