Make bookable unselectable when not available
When unavailable, the bookable is un-tappable and grayed out.

Big caveat: If you select the bookable, then change the time picker (or duration picker) to an interval when the bookable is unavailable, the selection is retained. This is not the desired behavior, but fixing it requires a fairly major change to the way we handle the form data. The form data needs to be held in the Redux state, and then validated in `mapStateToProps`, perhaps via some selector function.

If the user goes ahead and clicks “Bookit”, the server will, of course, return a 409, and then the existing behavior kicks in and we show the user a helpful message.